feat: refactor according to joshua's form for easier merge
