+++ src/components/App/App.tsx
@@ -11,7 +11,16 @@ const App: React.FC = () => {
           element={
             <div className="w-full max-w-[467px] mx-auto">
               {/* Wrapper for Form: The inner div ("w-full max-w-[467px] mx-auto") sets the maximum width and centers the form within that constrained width.*/}
-              <Form submit={() => console.log("Form submitted!")} />
+              <Form
+                submit={(data) =>
+                  console.log("Form submitted with data:", data)
+                }
+              >
+                <Form.TextAreaInput
+                  name="agreeToTerms"
+                  placeholder="Give us some details about your event"
+                />
+              </Form>
             </div>
           }
         />
