+++ src/components/Form/Form.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { useForm, FormProvider } from "react-hook-form";
+import { useForm, FormProvider, useFormContext } from "react-hook-form";
 import submitButton from "../../images/submitbutton.svg";
-import TextAreaInput from "../TextAreaInput/TextAreaInput";
 
 export type FormValues = {
   description: string;
@@ -10,11 +9,14 @@ export type FormValues = {
 
 interface FormProps {
   submit: (data: FormValues) => void;
+  children?: React.ReactNode;
 }
 
-const Form = ({ submit }: FormProps) => {
+const Form: React.FC<FormProps> & {
+  TextAreaInput: React.FC<TextAreaInputProps>;
+} = ({ submit, children }) => {
   const methods = useForm<FormValues>({
-    mode: "onBlur", // This will trigger validation on blur
+    mode: "onBlur", // Triggers validation on blur
   });
 
   const onSubmit = methods.handleSubmit((data) => {
@@ -29,15 +31,10 @@ const Form = ({ submit }: FormProps) => {
         className="bg-wheat p-[28px] pt-7 pb-10 rounded-[40px] max-w-[467px] w-full"
         onSubmit={onSubmit}
       >
-        <div className="flex flex-col justify-center items-center">
-          <div className="w-full mb-4">
-            <TextAreaInput
-              name="description"
-              placeholder="Give us some details about your event"
-            />
-          </div>
+        <div className="flex flex-col justify-center ">
+          {children}
 
-          <button type="submit" className="mt-4">
+          <button type="submit" className="flex mt-4 justify-center">
             <img
               className="w-[253px] h-[71px] opacity-100 hover:opacity-80 transition-opacity duration-300"
               src={submitButton}
@@ -49,5 +46,30 @@ const Form = ({ submit }: FormProps) => {
     </FormProvider>
   );
 };
+
+interface TextAreaInputProps {
+  name: keyof FormValues; // Restrict name to keys of FormValues
+  placeholder?: string;
+  error?: { message: string };
+}
+const TextAreaInput: React.FC<TextAreaInputProps> = ({ name, placeholder }) => {
+  const {
+    register,
+    formState: { errors },
+  } = useFormContext<FormValues>();
+  const errorMessage = errors[name]?.message as string | undefined;
+  return (
+    <div className="flex flex-col items-center">
+      <textarea
+        className="w-full h-[167px] rounded-[12px] p-3.5 bg-[#E9E8D5] mb-2.5 box-border 
+      text-base font-roboto leading-[19.2px] resize-none"
+        {...register(name, { required: "This field is required" })}
+        placeholder={placeholder}
+      />
+      {errorMessage && <p className="text-red">{errorMessage}</p>}
+    </div>
+  );
+};
+Form.TextAreaInput = TextAreaInput;
 
 export default Form;
